+++ hydra-node/src/Hydra/Network/Message.hs
@@ -4,12 +4,12 @@ module Hydra.Network.Message where
 
 import Hydra.Prelude
 
+import Cardano.Binary (serialize')
+import Cardano.Crypto.Util (SignableRepresentation, getSignableRepresentation)
 import Hydra.Crypto (Signature)
 import Hydra.Ledger (IsTx, UTxOType)
 import Hydra.Network (NodeId)
-import Hydra.Party (Party)
 import Hydra.Snapshot (Snapshot, SnapshotNumber)
-
 
 data Connectivity
   = Connected {nodeId :: NodeId}
@@ -17,13 +17,18 @@ data Connectivity
   deriving stock (Generic, Eq, Show)
   deriving anyclass (ToJSON, FromJSON)
 
-
--- NOTE(SN): Every message comes from a 'Party', we might want to move it out of
--- here into the 'NetworkEvent'
 data Message tx
-  = ReqTx {party :: Party, transaction :: tx}
-  | ReqSn {party :: Party, snapshotNumber :: SnapshotNumber, transactions :: [tx]}
-  | AckSn {party :: Party, signed :: Signature (Snapshot tx), snapshotNumber :: SnapshotNumber}
+  = ReqTx {transaction :: tx}
+  | ReqSn {snapshotNumber :: SnapshotNumber, transactions :: [tx]}
+  | -- NOTE: We remove the party from the ackSn but, it would make sense to put it
+    -- back as the signed snapshot is tied to the party and we should not
+    -- consider which party sent this message to validate this snapshot signature.
+    -- but currently we do not validate the snapshot signature itself, which is
+    -- a problem.
+    -- When we fix that, when we check the snapshot signature, that would be a
+    -- good idea to introduce the party in AckSn again or, maybe better, only
+    -- the verification key of the party.
+    AckSn {signed :: Signature (Snapshot tx), snapshotNumber :: SnapshotNumber}
   deriving stock (Generic, Eq, Show)
   deriving anyclass (ToJSON, FromJSON)
 
@@ -32,14 +37,17 @@ instance IsTx tx => Arbitrary (Message tx) where
 
 instance (ToCBOR tx, ToCBOR (UTxOType tx)) => ToCBOR (Message tx) where
   toCBOR = \case
-    ReqTx party tx -> toCBOR ("ReqTx" :: Text) <> toCBOR party <> toCBOR tx
-    ReqSn party sn txs -> toCBOR ("ReqSn" :: Text) <> toCBOR party <> toCBOR sn <> toCBOR txs
-    AckSn party sig sn -> toCBOR ("AckSn" :: Text) <> toCBOR party <> toCBOR sig <> toCBOR sn
+    ReqTx tx -> toCBOR ("ReqTx" :: Text) <> toCBOR tx
+    ReqSn sn txs -> toCBOR ("ReqSn" :: Text) <> toCBOR sn <> toCBOR txs
+    AckSn sig sn -> toCBOR ("AckSn" :: Text) <> toCBOR sig <> toCBOR sn
 
 instance (FromCBOR tx, FromCBOR (UTxOType tx)) => FromCBOR (Message tx) where
   fromCBOR =
     fromCBOR >>= \case
-      ("ReqTx" :: Text) -> ReqTx <$> fromCBOR <*> fromCBOR
-      "ReqSn" -> ReqSn <$> fromCBOR <*> fromCBOR <*> fromCBOR
-      "AckSn" -> AckSn <$> fromCBOR <*> fromCBOR <*> fromCBOR
+      ("ReqTx" :: Text) -> ReqTx <$> fromCBOR
+      "ReqSn" -> ReqSn <$> fromCBOR <*> fromCBOR
+      "AckSn" -> AckSn <$> fromCBOR <*> fromCBOR
       msg -> fail $ show msg <> " is not a proper CBOR-encoded Message"
+
+instance IsTx tx => SignableRepresentation (Message tx) where
+  getSignableRepresentation = serialize'