+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 As a minor extension, we also keep a semantic version for the `UNRELEASED`
 changes.
 
+## [0.12.0] - UNRELEASED
+
+- **BREAKING** Introduce authenticated network messages [#965](965).
+  - Peers will sign messages before broadcasting them to other peers,
+    and verify signature of received messages is from a known party
+    and of course valid.
+
 ## [0.11.0] - 2023-06-30
 
 This release contains breaking changes of the persistence and on-chain scripts